json読み込みでエラーが起きている

+++ Source/Framework/Definition/Constant/Graphics/Resource/Buffer/Constant/ConstantBufferConstant.h
@@ -4,5 +4,5 @@ namespace FWK::Constant
 {
 	inline constexpr UINT64 k_constantBufferAlignment = 256ULL;
 
-	inline constexpr UINT k_invalidCreateConstantBufferNUM = 0U;
+	inline constexpr UINT64 k_invalidCreateConstantBufferNUM = 0ULL;
 }
